blood

+++ solo/data/classification_dataloader.py
@@ -29,7 +29,6 @@ from torch.utils.data import DataLoader, Dataset
 from torchvision import transforms
 from torchvision.datasets import STL10, ImageFolder
 import medmnist
-import numpy as np
 
 try:
     from solo.data.h5_dataset import H5Dataset
@@ -258,7 +257,7 @@ def prepare_datasets(
             size=128
         )
         for i in range(len(train_dataset[0][1])):
-            train_dataset[0][1][i] = train_dataset[0][1][i].astype(np.float16)
+            train_dataset[0][1][i] = train_dataset[0][1][i].to(dtype=torch.float32)
         DatasetClass = vars(medmnist)[dataset]
         val_dataset = DatasetClass(
             root=val_data_path,
@@ -268,7 +267,7 @@ def prepare_datasets(
             size=128
         )
         for i in range(len(train_dataset[0][1])):
-            val_dataset[0][1][i] = val_dataset[0][1][i].astype(np.float16)
+            val_dataset[0][1][i] = val_dataset[0][1][i].to(dtype=torch.float32)
 
     elif dataset == "stl10":
         train_dataset = STL10(
